agents/sapper/requirements.txt
# Sapper requirements
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Messaging
pika==1.3.2
aio-pika==9.4.1

# SSH/Network
asyncssh==2.14.2
paramiko==3.4.0

# HTTP clients
httpx==0.25.2
requests==2.31.0

# Caching
redis==5.0.1
hiredis==2.2.3

# Logging & Monitoring
python-json-logger==2.0.7
prometheus-client==0.19.0

# Utilities
PyYAML==6.0.1
tenacity==8.2.3

# Email
aiosmtplib>=3.0,<4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
